+++ src/app.module.ts
@@ -3,7 +3,7 @@ import { ConfigModule } from '@nestjs/config';
 import { AppController } from './app.controller';
 import { AppService } from './app.service';
 import { CatsModule } from './cats/cats.module';
-import { AuthModule } from './auth/auth.module';
+import { AuthModule } from './authn/auth.module';
 import { UsersModule } from './users/users.module';
 import { AuthzService } from './authz/authz.service';
 import { AuthzModule } from './authz/authz.module';
